==> Code_PhD_Ch04/04b_Constant_Management_Cost.R <==
# Asign constant management budget for each year 
CIdata2001_MANAGEMENT_BUDGET <- vector()
CIdata2003_MANAGEMENT_BUDGET <- vector()
CIdata2005_MANAGEMENT_BUDGET <- vector()
CIdata2007_MANAGEMENT_BUDGET <- vector()
CIdata2009_MANAGEMENT_BUDGET <- vector()
CIdata2011_MANAGEMENT_BUDGET <- vector()
CIdata2013_MANAGEMENT_BUDGET <- vector()
CIdata2015_MANAGEMENT_BUDGET <- vector()

constant_management_budget <- 6000

for (i in 1:100){
  CIdata2001_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2003_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2005_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2007_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2009_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2011_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2013_MANAGEMENT_BUDGET[i] <- constant_management_budget
  CIdata2015_MANAGEMENT_BUDGET[i] <- constant_management_budget
}
